+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.11.0] - 2018-12-13
+### Changed
+- Do not set empty SID from ORACLE_SID/TWO_TASK enviroment variables, leave it to ODPI.
+
+### Added
+- Allow PRELIM authentication to allow Startup and Shutdown.
+
 ## [2.10.1] - 2018-11-23
 ### Changed
 - Don't call SET TRANSACTION if not really needed in BeginTx - if the isolation level hasn't changed.